requirements.txt
boto3==1.40.35
botocore==1.40.35
jmespath==1.0.1
numpy==2.3.3
pandas==2.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
s3transfer==0.14.0
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
